refactor(internal): Introduce js-utils.js
