+++ packages/internal/src/index.js
@@ -5,6 +5,7 @@
 export * from './config.js';
 export * from './debug.js';
 export * from './errors.js';
+export * from './js-utils.js';
 export * from './method-tools.js';
 export * from './ses-utils.js';
 export * from './typeCheck.js';
