+++ packages/internal/test/snapshots/exports.test.js.md
@@ -9,7 +9,6 @@ Generated by [AVA](https://avajs.dev).
 > Snapshot 1
 
     [
-      'BASIS_POINTS',
       'BridgeId',
       'CosmosInitKeyToBridgeId',
       'NonNullish',
