+++ packages/internal/test/utils.test.js
@@ -2,14 +2,13 @@
 import test from 'ava';
 
 import { Far } from '@endo/far';
+import { deepMapObject, makeMeasureSeconds } from '../src/js-utils.js';
 import {
-  makeMeasureSeconds,
   assertAllDefined,
   whileTrue,
   untilTrue,
   forever,
   deeplyFulfilledObject,
-  deepMapObject,
   synchronizedTee,
 } from '../src/ses-utils.js';
